prefNum を扱うロジックを usePrefNum.ts カスタムフックに分割した
